+++ core/components/romanesco/elements/snippets/06_formulas/f_presentation/sliderloadassets.snippet.php
@@ -65,7 +65,7 @@ foreach ($behaviour as $option) {
 $clickable = ($pagination == 'bullets') ? 'true' : 'false';
 
 // Effects
-$effects = array(
+$transitionEffects = array(
     'fade' => '
         fadeEffect: {
             crossFade: true
@@ -78,6 +78,7 @@ $effects = array(
         },
     ',
     'flip' => '
+        grabCursor: true,
         flipEffect: {
             rotate: 30,
             slideShadows: false,
@@ -88,7 +89,39 @@ $effects = array(
             slideShadows: false,
         },
     ',
+    'parallax' => '
+        speed: 600,
+        parallax: true,
+    ',
+    'photocopy' => '
+        creativeEffect: {
+            prev: {
+                shadow: false,
+                translate: ["-20%", 0, -1],
+            },
+            next: {
+                translate: ["100%", 0, 0],
+            },
+        },
+    ',
+    'escalator' => '
+        creativeEffect: {
+            prev: {
+                shadow: true,
+                translate: [0, 0, -400],
+            },
+            next: {
+                translate: ["100%", 0, 0],
+            },
+        },
+    ',
 );
+$effects = $transitionEffects[$transition];
+
+// Rename custom (creative) effects
+if (in_array($transition, ['photocopy','escalator'])) {
+    $transition = 'creative';
+}
 
 // Responsive
 if ($responsive) {
@@ -194,7 +227,7 @@ $modx->regClientHTMLBlock($modx->getChunk($tpl, array(
     'pagination' => $pagination ?? '',
     'clickable' => $clickable,
     'breakpoints' => $breakpoints ?? '',
-    'effects' => $effects[$transition] ?? '',
+    'effects' => $effects ?? '',
     'init_lightbox' => $initLightbox ?? '',
 )));